Move Monit and Dash to Extra and Clean use file
