+++ lib/API.js
@@ -1574,80 +1574,6 @@ API.prototype.describe = function(pm2_id, cb) {
 };
 
 /**
- * Description
- * @method monit
- * @return
- */
-API.prototype.monit = function(cb) {
-  var that = this;
-
-  var Monit = require('./API/Monit');
-
-  if (cb) return cb(new Error('Monit cant be called programmatically'));
-
-  Monit.init();
-
-  function launchMonitor() {
-    that.Client.executeRemote('getMonitorData', {}, function(err, list) {
-      if (err) {
-        console.error('Error retrieving process list: ' + err);
-        that.exitCli(conf.ERROR_EXIT);
-      }
-
-      Monit.refresh(list);
-
-      setTimeout(function() {
-        launchMonitor();
-      }, 400);
-    });
-  }
-
-  launchMonitor();
-};
-
-/**
- * Description
- * @method dashboard
- * @return
- */
-API.prototype.dashboard = function(cb) {
-  var that = this;
-
-  var Dashboard = require('./API/Dashboard');
-
-  if (cb) return cb(new Error('Dashboard cant be called programmatically'));
-
-  Dashboard.init();
-
-  this.Client.launchBus(function (err, bus) {
-    if (err) {
-        console.error('Error launchBus: ' + err);
-        that.exitCli(conf.ERROR_EXIT);
-    }
-    bus.on('log:*', function(type, data) {
-      Dashboard.log(type, data);
-    })
-  });
-
-  function launchDashboard() {
-    that.Client.executeRemote('getMonitorData', {}, function(err, list) {
-      if (err) {
-        console.error('Error retrieving process list: ' + err);
-        that.exitCli(conf.ERROR_EXIT);
-      }
-
-      Dashboard.refresh(list);
-
-      setTimeout(function() {
-        launchDashboard();
-      }, 400);
-    });
-  }
-
-  launchDashboard();
-};
-
-/**
  * API method to perform a deep update of PM2
  * @method deepUpdate
  */
