+++ lib/API/Extra.js
@@ -327,4 +327,78 @@ module.exports = function(CLI) {
     Common.printOut('File %s generated', path.join(pwd, f_name));
     that.exitCli(cst.SUCCESS_EXIT);
   };
+
+  /**
+   * Description
+   * @method monit
+   * @return
+   */
+  CLI.prototype.monit = function(cb) {
+    var that = this;
+
+    var Monit = require('./Monit');
+
+    if (cb) return cb(new Error('Monit cant be called programmatically'));
+
+    Monit.init();
+
+    function launchMonitor() {
+      that.Client.executeRemote('getMonitorData', {}, function(err, list) {
+        if (err) {
+          console.error('Error retrieving process list: ' + err);
+          that.exitCli(conf.ERROR_EXIT);
+        }
+
+        Monit.refresh(list);
+
+        setTimeout(function() {
+          launchMonitor();
+        }, 400);
+      });
+    }
+
+    launchMonitor();
+  };
+
+  /**
+   * Description
+   * @method dashboard
+   * @return
+   */
+  CLI.prototype.dashboard = function(cb) {
+    var that = this;
+
+    var Dashboard = require('./Dashboard');
+
+    if (cb) return cb(new Error('Dashboard cant be called programmatically'));
+
+    Dashboard.init();
+
+    this.Client.launchBus(function (err, bus) {
+      if (err) {
+          console.error('Error launchBus: ' + err);
+          that.exitCli(conf.ERROR_EXIT);
+      }
+      bus.on('log:*', function(type, data) {
+        Dashboard.log(type, data);
+      })
+    });
+
+    function launchDashboard() {
+      that.Client.executeRemote('getMonitorData', {}, function(err, list) {
+        if (err) {
+          console.error('Error retrieving process list: ' + err);
+          that.exitCli(conf.ERROR_EXIT);
+        }
+
+        Dashboard.refresh(list);
+
+        setTimeout(function() {
+          launchDashboard();
+        }, 400);
+      });
+    }
+
+    launchDashboard();
+  };
 };
